SD Chute Pooping For Test Rocket V.1
We can read the altitude and write it to the SD card. It also tells us
the direction that we are going (up or down). And it will also tell us
when it would pop the parachute. This data will all be saved on the SD
card.

Things to do:
-make it so that it creates a differently named file each time
-add accelerometer stuff